Refactor : 좌측 네비게이션 컴포넌트 분리
